+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## 1.1.4 (2022-)
+
 - `arrTo1dArray`: Add `arrTo1dArray` method
 
 ## 1.1.3 (2022-10-23)